Add option to set alternative pacman command
This patch introduces the ability to pick a different pacman command to
run when searching for installed packages and installing the meta-package.
This is mainly useful for using aur helpers.

The option to provide an alternative pacman command is -p or --pacman,
followed by the command name or its path.

With the use of aur helpers in mind, it also adds the following options:
    - --not-as-root: Run installation command as unpriviledged user
    - --depends-first: Run a separate instance of pacman (or the
      provided alternative) to install the dependencies before
      installing the meta-package itself.

Signed-off-by: Mateus Auler <mateusauler@protonmail.com>